backend/requirements.txt
1337x==1.2.3
appdirs==1.4.4
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cattrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.0rc8
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
requests-cache==0.9.6
six==1.16.0
soupsieve==2.3.2.post1
url-normalize==1.4.3
urllib3==1.26.12
Werkzeug==2.2.2
